+++ lib/appmetrica_sdk.dart
@@ -5,6 +5,7 @@
 library appmetrica_sdk;
 
 import 'dart:async';
+import 'dart:io';
 import 'package:flutter/services.dart';
 
 part 'src/appmetrica_sdk.dart';